Remove babel compilation, use native Node module support.
Tape's runner always tries to `require` the test files (which isn't
allowed with Node's module stuff), so I switched it to just run the
singular test file for now. If we add more test files we can figure it
out then (or maybe it will be fixed by then).

The other necessary change was around the JSON imports (which aren't a
thing in modules either). I changed these into extremely simple modules
(since it seems like that will be the least-break-y way to change that
vs `readFile` which may break webpack/other build solutions?).